Refactored minutes parsing, sanitizing, and formatting for PDF into utility function
- The utility function takes minutesState as input and outputs an object containing pdfReadyMinutes, ready for PDF generation.
- Functionality includes preprocessing checkboxes, converting Markdown to HTML, and sanitizing the HTML to allow only preferred tags and attributes. It also transforms and styles tags that can't be directly rendered by react-pdf-html.
- Designed to support future features, such as a DownloadPDF button.
